remove vx depedencies
